requirements-ui.txt
# User-facing PoC dashboard
# pip install -r requirements-ui.txt
streamlit>=1.28
plotly>=5.18
fpdf2>=2.7
